// src/generated/models/Strategy_strategiesModel.ts
export const Strategy_strategiesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Strategy_strategiesstatecode = keyof typeof Strategy_strategiesstatecode;
export const Strategy_strategiesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Strategy_strategiesstatuscode = keyof typeof Strategy_strategiesstatuscode;
export const Strategy_strategiesstf_revisionstatus = {
  1: 'Draft',
  2: 'Submitted',
  3: 'Under Review',
  4: 'Changes Requested',
  5: 'Approved (Locked)',
  6: 'Rejected',
  7: 'Reopened'
} as const;
export type Strategy_strategiesstf_revisionstatus = keyof typeof Strategy_strategiesstf_revisionstatus;
export const Strategy_strategiesstf_strategytrack = {
  1: 'Operational',
  2: 'Service'
} as const;
export type Strategy_strategiesstf_strategytrack = keyof typeof Strategy_strategiesstf_strategytrack;
export const Strategy_strategiesstrategy_complexity = {
  620930000: 'High Complexity & High Time',
  620930001: 'Low Complexity & Low Time',
  620930002: 'Low Complexity & High Time',
  620930003: 'High Complexity & Low Time'
} as const;
export type Strategy_strategiesstrategy_complexity = keyof typeof Strategy_strategiesstrategy_complexity;
export const Strategy_strategiesstrategy_financialconfirmation = {
  620930000: 'Pending',
  620930001: 'Confermed'
} as const;
export type Strategy_strategiesstrategy_financialconfirmation = keyof typeof Strategy_strategiesstrategy_financialconfirmation;
export const Strategy_strategiesstrategy_strategylevel = {
  620930000: 'New',
  620930001: 'continuing'
} as const;
export type Strategy_strategiesstrategy_strategylevel = keyof typeof Strategy_strategiesstrategy_strategylevel;
export const Strategy_strategiesstrategy_strategystatus = {
  620930000: 'Strategy Design Check',
  620930001: 'Strategy Cascading Check (Defined moves)',
  989230001: 'Execution Readiness Check (pre-execution)',
  989230002: 'Execution & Governance Check',
  989230003: 'Strategy Performance Monitoring Check'
} as const;
export type Strategy_strategiesstrategy_strategystatus = keyof typeof Strategy_strategiesstrategy_strategystatus;
export const Strategy_strategiesstrategy_strategytype = {
  989230000: 'Department Strategy',
  989230001: 'Market Strategy',
  989230002: 'Specialty Strategy',
  989230003: 'Service Strategy'
} as const;
export type Strategy_strategiesstrategy_strategytype = keyof typeof Strategy_strategiesstrategy_strategytype;

export interface Strategy_strategiesBase {
  "cr18c_BusinessUnit@odata.bind"?: string;
  "cr18c_Department@odata.bind": string;
  "cr18c_Process@odata.bind"?: string;
  cr18c_specialty?: string;
  "cr18c_Subprocess@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Strategy_strategiesstatecode;
  statuscode?: Strategy_strategiesstatuscode;
  "stf_ApprovedBy@odata.bind"?: string;
  stf_approvedon?: string;
  "stf_ObjectiveDepartment@odata.bind"?: string;
  stf_revisionstatus?: Strategy_strategiesstf_revisionstatus;
  stf_strategytrack: Strategy_strategiesstf_strategytrack;
  "stf_SupportedDepartment@odata.bind"?: string;
  "stf_SupportedStrategy@odata.bind"?: string;
  "stf_SupportiveFunction@odata.bind"?: string;
  strategy_baseline?: number;
  "strategy_Company@odata.bind": string;
  strategy_complexity: Strategy_strategiesstrategy_complexity;
  strategy_conversationid?: string;
  strategy_enddate: string;
  strategy_financialconfirmation?: Strategy_strategiesstrategy_financialconfirmation;
  "strategy_Function@odata.bind": string;
  strategy_implementationconfidence: number;
  "strategy_KPI@odata.bind": string;
  strategy_kpiactual?: number;
  strategy_kpitarget?: number;
  strategy_newcolumn: string;
  "strategy_Region@odata.bind": string;
  "strategy_Section@odata.bind"?: string;
  strategy_startdate: string;
  strategy_strategydescription?: string;
  strategy_strategyid: string;
  strategy_strategylevel: Strategy_strategiesstrategy_strategylevel;
  strategy_strategystatus: Strategy_strategiesstrategy_strategystatus;
  strategy_strategytype: Strategy_strategiesstrategy_strategytype;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}